[loop_hook_redraw.c] norminette via extraction to raytrace.c
