+++ includes/loop_hook.h
@@ -6,13 +6,14 @@
 /*   By: oadhesiv <secondfry+school21@gmail.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/06 17:50:14 by oadhesiv          #+#    #+#             */
-/*   Updated: 2021/06/27 16:43:02 by oadhesiv         ###   ########.fr       */
+/*   Updated: 2021/07/17 14:26:43 by oadhesiv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LOOP_HOOK_H
 # define LOOP_HOOK_H
 
+# include <pthread.h>
 # include <math.h>
 # include "init_mlx.h"
 # include "color.h"
@@ -20,6 +21,17 @@
 # include "matrix_utils.h"
 # include "intersection.h"
 # include "vector.h"
+# include "raytrace.h"
+
+typedef struct s_params {
+	t_rtv	*rtv;
+	short	ya;
+	short	yb;
+	short	xa;
+	short	xb;
+}				t_params;
+
+# define THREAD_COUNT 8
 
 int		loop_hook(t_rtv *rtv);
 int		loop_destroy_hook(t_rtv *rtv);
